:rocket:feat: Added Result Activity style and gasoline or ethanol calculator

+++ app/src/main/java/com/example/gasolineorethanol/ui/activity/GasolineOrEthanolFormActivity.kt
@@ -19,8 +19,8 @@ class GasolineOrEthanolFormActivity : AppCompatActivity() {
         val resultButton = findViewById<Button>(R.id.form_result_button)
         resultButton.setOnClickListener {
             if (gasolineField.text.isNotEmpty() && ethanolField.text.isNotEmpty()) {
-                fuelsData.setGasolinePrice(BigDecimal(gasolineField.text.toString()))
-                fuelsData.setEthanolPrice(BigDecimal(ethanolField.text.toString()))
+                fuelsData.setGasolinePrice(gasolineField.text.toString().toDouble())
+                fuelsData.setEthanolPrice(ethanolField.text.toString().toDouble())
                 val intent = Intent(this, ResultActivity::class.java)
                 startActivity(intent)
             }
